ref(linter): move tests to lint files
