+++ linter/src/rules/bad_drop_database.rs
@@ -18,3 +18,18 @@ pub fn ban_drop_database(tree: &[RootStmt]) -> Vec<RuleViolation> {
     }
     errs
 }
+
+#[cfg(test)]
+mod test_rules {
+    use crate::check_sql;
+    use insta::assert_debug_snapshot;
+    #[test]
+    fn test_ban_drop_database() {
+        let sql = r#"
+DROP DATABASE "table_name";
+DROP DATABASE IF EXISTS "table_name";
+DROP DATABASE IF EXISTS "table_name"
+        "#;
+        assert_debug_snapshot!(check_sql(sql, &[]));
+    }
+}
