+++ linter/src/rules/changing_column_type.rs
@@ -21,3 +21,33 @@ pub fn changing_column_type(tree: &[RootStmt]) -> Vec<RuleViolation> {
     }
     errs
 }
+
+#[cfg(test)]
+mod test_rules {
+    use crate::check_sql;
+    use insta::assert_debug_snapshot;
+    #[test]
+    fn test_changing_field_type() {
+        let bad_sql = r#"
+BEGIN;
+--
+-- Alter field edits on recipe
+--
+ALTER TABLE "core_recipe" ALTER COLUMN "edits" TYPE text USING "edits"::text;
+COMMIT;
+        "#;
+        assert_debug_snapshot!(check_sql(bad_sql, &[]));
+
+        let bad_sql = r#"
+BEGIN;
+--
+-- Alter field foo on recipe
+--
+ALTER TABLE "core_recipe" ALTER COLUMN "foo" TYPE varchar(255) USING "foo"::varchar(255);
+ALTER TABLE "core_recipe" ALTER COLUMN "foo" TYPE text USING "foo"::text;
+COMMIT;
+        "#;
+
+        assert_debug_snapshot!(check_sql(bad_sql, &[]));
+    }
+}
