+++ linter/src/rules/disallow_unique_constraint.rs
@@ -33,3 +33,47 @@ pub fn disallow_unique_constraint(tree: &[RootStmt]) -> Vec<RuleViolation> {
     }
     errs
 }
+
+#[cfg(test)]
+mod test_rules {
+    use crate::check_sql;
+    use insta::assert_debug_snapshot;
+
+    /// ```sql
+    /// -- instead of
+    /// ALTER TABLE table_name ADD CONSTRAINT field_name_constraint UNIQUE (field_name);
+    /// -- also works with PRIMARY KEY
+    /// -- use:
+    /// -- To recreate a primary key constraint, without blocking updates while the index is rebuilt:
+    /// CREATE UNIQUE INDEX CONCURRENTLY dist_id_temp_idx ON distributors (dist_id);
+    /// ALTER TABLE distributors DROP CONSTRAINT distributors_pkey,
+    /// ADD CONSTRAINT distributors_pkey PRIMARY KEY USING INDEX dist_id_temp_idx;
+    /// ```
+    #[test]
+    fn test_adding_unique_constraint() {
+        let bad_sql = r#"
+ALTER TABLE table_name ADD CONSTRAINT field_name_constraint UNIQUE (field_name);
+   "#;
+
+        let ok_sql = r#"
+CREATE UNIQUE INDEX CONCURRENTLY dist_id_temp_idx ON distributors (dist_id);
+ALTER TABLE distributors DROP CONSTRAINT distributors_pkey,
+ADD CONSTRAINT distributors_pkey PRIMARY KEY USING INDEX dist_id_temp_idx;
+   "#;
+
+        assert_debug_snapshot!(check_sql(bad_sql, &["prefer-robust-stmts".into()]));
+        assert_debug_snapshot!(check_sql(ok_sql, &["prefer-robust-stmts".into()]));
+    }
+
+    /// Creating a UNQIUE constraint from an existing index should be considered
+    /// safe
+    #[test]
+    fn test_unique_constraint_ok() {
+        let sql = r#"
+CREATE UNIQUE INDEX CONCURRENTLY "legacy_questiongrouppg_mongo_id_1f8f47d9_uniq_idx" 
+    ON "legacy_questiongrouppg" ("mongo_id");
+ALTER TABLE "legacy_questiongrouppg" ADD CONSTRAINT "legacy_questiongrouppg_mongo_id_1f8f47d9_uniq" UNIQUE USING INDEX "legacy_questiongrouppg_mongo_id_1f8f47d9_uniq_idx";
+        "#;
+        assert_eq!(check_sql(sql, &["prefer-robust-stmts".into()]), Ok(vec![]));
+    }
+}
