+++ linter/src/rules/require_concurrent_index_creation.rs
@@ -23,3 +23,31 @@ pub fn require_concurrent_index_creation(tree: &[RootStmt]) -> Vec<RuleViolation
     }
     errs
 }
+
+#[cfg(test)]
+mod test_rules {
+    use crate::check_sql;
+    use insta::assert_debug_snapshot;
+
+    /// ```sql
+    /// -- instead of
+    /// CREATE INDEX "field_name_idx" ON "table_name" ("field_name");
+    /// -- use CONCURRENTLY
+    /// CREATE INDEX CONCURRENTLY "field_name_idx" ON "table_name" ("field_name");
+    /// ```
+    #[test]
+    fn test_adding_index_non_concurrently() {
+        let bad_sql = r#"
+  -- instead of
+  CREATE INDEX "field_name_idx" ON "table_name" ("field_name");
+  "#;
+
+        assert_debug_snapshot!(check_sql(bad_sql, &["prefer-robust-stmts".into()]));
+
+        let ok_sql = r#"
+  -- use CONCURRENTLY
+  CREATE INDEX CONCURRENTLY "field_name_idx" ON "table_name" ("field_name");
+  "#;
+        assert_debug_snapshot!(check_sql(ok_sql, &["prefer-robust-stmts".into()]));
+    }
+}
